requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==1.8.1
